Use hardware volume keys from anywhere inside Kore
* Move dispatchKeyEvent callback for volume control to the BaseActivity to allow all activities to receive hardware button events

resolves xbmc/Kore#235
