+++ app/src/main/java/org/xbmc/kore/ui/sections/remote/RemoteActivity.java
@@ -86,11 +86,6 @@ public class RemoteActivity extends BaseActivity
     private static final int PLAYLIST_FRAGMENT_ID = 3;
 
     /**
-     * Host manager singleton
-     */
-    private HostManager hostManager = null;
-
-    /**
      * To register for observing host events
      */
     private HostConnectionObserver hostConnectionObserver;
@@ -111,18 +106,6 @@ public class RemoteActivity extends BaseActivity
 
         setContentView(R.layout.activity_remote);
         ButterKnife.inject(this);
-
-        hostManager = HostManager.getInstance(this);
-
-        // Check if we have any hosts setup
-        if (hostManager.getHostInfo() == null) {
-            final Intent intent = new Intent(this, AddHostActivity.class);
-            intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-            intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            startActivity(intent);
-            finish();
-            return;
-        }
 
         // Set up the drawer.
         navigationDrawerFragment = (NavigationDrawerFragment) getSupportFragmentManager()
@@ -199,40 +182,6 @@ public class RemoteActivity extends BaseActivity
         super.onPause();
         if (hostConnectionObserver != null) hostConnectionObserver.unregisterPlayerObserver(this);
         hostConnectionObserver = null;
-    }
-
-    /**
-     * Override hardware volume keys and send to Kodi
-     */
-    @Override
-    public boolean dispatchKeyEvent(KeyEvent event) {
-        // Check whether we should intercept this
-        boolean useVolumeKeys = PreferenceManager
-                .getDefaultSharedPreferences(this)
-                .getBoolean(Settings.KEY_PREF_USE_HARDWARE_VOLUME_KEYS,
-                        Settings.DEFAULT_PREF_USE_HARDWARE_VOLUME_KEYS);
-        if (useVolumeKeys) {
-            int action = event.getAction();
-            int keyCode = event.getKeyCode();
-            switch (keyCode) {
-                case KeyEvent.KEYCODE_VOLUME_UP:
-                    if (action == KeyEvent.ACTION_DOWN) {
-                        new Application
-                                .SetVolume(GlobalType.IncrementDecrement.INCREMENT)
-                                .execute(hostManager.getConnection(), null, null);
-                    }
-                    return true;
-                case KeyEvent.KEYCODE_VOLUME_DOWN:
-                    if (action == KeyEvent.ACTION_DOWN) {
-                        new Application
-                                .SetVolume(GlobalType.IncrementDecrement.DECREMENT)
-                                .execute(hostManager.getConnection(), null, null);
-                    }
-                    return true;
-            }
-        }
-
-        return super.dispatchKeyEvent(event);
     }
 
     @Override
